+++ src/components/screen/composeGame.js
@@ -3,25 +3,34 @@ export default (state) => ({
   init() {
     console.log('init running from composeGame');
   },
-  foo(playerReady) {
-    const isPlayerReady = this.playersReady.some((player) => player === playerReady);
-    if (!isPlayerReady) {
-      this.playersReady.push(playerReady);
+  isGameReady(player, isReady) {
+    if (this.mode) {
+      // If human vs human
+      const index = this.playersReady.indexOf(player);
+      if (player && isReady !== undefined) {
+        if (isReady) {
+          this.playersReady.push(player);
+        } else {
+          this.playersReady.splice(index, 1);
+        }
+      } else if (index > -1) {
+        this.playersReady.splice(index, 1);
+      }
+
+      if (this.playersReady.length === 2 && this.playBtn.classList.contains('inactive')) {
+        this.playBtn.classList.remove('inactive');
+      } else {
+        this.playBtn.classList.add('inactive');
+      }
     } else {
-      const index = this.playersReady.indexOf(isPlayerReady);
-      this.playersReady.splice(index, 1);
-    }
-    console.log(this.playersReady);
-    if (this.playersReady.length === 2 && this.startBtn.classList.contains('inactive')) {
-      this.startBtn.classList.remove('inactive');
-    } else {
-      this.startBtn.classList.add('inactive');
+      if (this.playerTwoContainer.classList.contains('inactive')) {
+        this.playerTwoContainer.classList.remove('inactive');
+      } else {
+        this.playerTwoContainer.classList.add('inactive');
+      }
     }
   },
-  start(e) {
-    // Set this.gameReady to true
-    // Publish something...?
-    // Reveal player two's board
+  play(e) {
     if (!this.mode) {
       this.game.playerTwo.board.placeShipsRandom();
     }