+++ src/components/screen/screenController.js
@@ -27,8 +27,8 @@ export default (mode) => {
       };
       pubSub.publish('notify', 'Place ships');
       this.updateGameState(composeGame);
-      this.start = this.start.bind(this);
-      this.foo = this.foo.bind(this);
+      this.play = this.play.bind(this);
+      this.isGameReady = this.isGameReady.bind(this);
     },
     updateGameState(callback) {
       Object.assign(this, callback());
@@ -42,14 +42,13 @@ export default (mode) => {
       this.playerTwoBoard = element.querySelector('.player_two > .board');
       this.playerOneHeader = element.querySelector('.player_one > h4');
       this.playerTwoHeader = element.querySelector('.player_two > h4');
-      this.startBtn = element.querySelector('.game_start_btn');
-      console.log(this.startBtn);
+      this.playBtn = element.querySelector('.play_btn');
     },
     bindEvents() {
       if (!this.gameReady) {
         // if (!this.mode) {
-        this.startBtn.addEventListener('click', this.start);
-        pubSub.subscribe('playerReady', this.foo);
+        this.playBtn.addEventListener('click', this.play);
+        pubSub.subscribe('playerReady', this.isGameReady);
         // }
       }
 
@@ -73,18 +72,18 @@ export default (mode) => {
       const playerTwoContainer = createElement('div');
       const playerOneHeader = createElement('h4');
       const playerTwoHeader = createElement('h4');
-      const gameStartContainer = createElement('div');
-      const gameStartBtn = createElement('button');
-      const gameStartBtnText = createElement('span');
-      gameStartBtnText.textContent = 'Play';
+      const gamePlayContainer = createElement('div');
+      const gamePlayBtn = createElement('button');
+      const gamePlayBtnText = createElement('span');
+      gamePlayBtnText.textContent = 'Play';
       gameContainer.id = 'game_container';
       boardsContainer.id = 'boards_container';
       playerOneContainer.classList.add('player_one');
       playerTwoContainer.classList.add('player_two');
       playerOneHeader.textContent = `Player one's grid`;
       playerTwoHeader.textContent = `Player two's grid`;
-      gameStartContainer.classList.add('game_start');
-      gameStartBtn.classList.add('game_start_btn', 'inactive');
+      gamePlayContainer.classList.add('game_play');
+      gamePlayBtn.classList.add('play_btn');
       // Renders players' boards
       playerOneContainer.appendChild(board(this.boards.playerOne));
       playerTwoContainer.appendChild(board(this.boards.playerTwo));
@@ -92,16 +91,18 @@ export default (mode) => {
       playerTwoContainer.appendChild(playerTwoHeader);
       boardsContainer.appendChild(playerOneContainer);
       boardsContainer.appendChild(playerTwoContainer);
-      gameStartBtn.appendChild(gameStartBtnText);
-      gameStartContainer.appendChild(gameStartBtn);
+      gamePlayBtn.appendChild(gamePlayBtnText);
+      gamePlayContainer.appendChild(gamePlayBtn);
       if (!this.gameReady) {
-        playerOneContainer.appendChild(port('player_one', this.game));
+        playerOneContainer.appendChild(port('player_one', this.game, this.mode));
         if (this.mode) {
-          playerTwoContainer.appendChild(port('player_two', this.game));
+          playerTwoContainer.appendChild(port('player_two', this.game, this.mode));
+          gamePlayBtn.classList.add('inactive');
         } else {
+          playerTwoContainer.classList.add('inactive');
           playerTwoContainer.classList.add('wait');
         }
-        playerTwoContainer.appendChild(gameStartContainer);
+        playerTwoContainer.appendChild(gamePlayContainer);
       }
 
       gameContainer.appendChild(boardsContainer);